refactor: rename and relocate misplaced tests in sessions modules
- Rename persistence lifecycle tests in create.rs to reflect what they
  actually test (save/load/remove cycles, not destroy behavior)
- Move destroy_session tests from create.rs to destroy.rs where they
  belong
